requirements.txt
PyJWT>=2.0.0
aws-lambda-powertools>=2.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
boto3==1.28.44
python-dotenv==1.0.0
pytest>=7.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6